# message_service/docs/swagger.yaml
basePath: /
definitions:
...
  controllers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  controllers.LoginResponse:
    properties:
      token:
        type: string
      user:
        properties:
          id:
            type: integer
          role:
            type: integer
        type: object
    type: object
  controllers.LogoutResponse:
    properties:
      message:
        type: string
    type: object
...
  controllers.RegisterResponse:
    properties:
      message:
        type: string
      role:
        type: integer
      user_id:
        type: integer
    type: object
...
  models.Role:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.Role'
        description: Связь с моделью Role
      role_id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Этот эндпоинт выполняет аутентификацию пользователя и возвращает
        JWT токен.
      parameters:
      - description: Информация для аутентификации
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Аутентификация пользователя
      tags:
      - Пользователи
  /logout:
    post:
      consumes:
      - application/json
      description: Этот эндпоинт завершает текущую сессию пользователя.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.LogoutResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Выход пользователя
      tags:
      - Пользователи
...
  /register:
    post:
      consumes:
      - application/json
      description: Этот эндпоинт регистрирует нового пользователя с уникальными именем
        пользователя и email.
      parameters:
      - description: Информация о пользователе
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Регистрация пользователя
      tags:
      - Пользователи
...
swagger: "2.0"